+++ vrp-core/src/models/examples.rs
@@ -1,5 +1,5 @@
 use crate::construction::constraints::{ConstraintPipeline, TransportConstraintModule};
-use crate::models::common::{TimeInterval, TimeSpan, TimeWindow};
+use crate::models::common::{Dimensions, IdDimension, TimeInterval, TimeSpan, TimeWindow};
 use crate::models::problem::*;
 use crate::models::Problem;
 use std::sync::Arc;
@@ -39,10 +39,12 @@ fn create_example_fleet() -> Arc<Fleet> {
         dimens: Default::default(),
         details: vec![],
     })];
+    let mut vehicle_dimens = Dimensions::new();
+    vehicle_dimens.set_id("v1");
     let vehicles = vec![Arc::new(Vehicle {
         profile: 0,
         costs: Costs { fixed: 0., per_distance: 1., per_driving_time: 0., per_waiting_time: 0., per_service_time: 0. },
-        dimens: Default::default(),
+        dimens: vehicle_dimens,
         details: vec![VehicleDetail {
             start: Some(VehiclePlace { location: 0, time: TimeInterval::default() }),
             end: None,